# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users":{"post":{"operationId":"UserController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSignIn"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/refresh":{"get":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/kategori":{"post":{"operationId":"KategoriController_create","summary":"","description":"Create a new Kategori Record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateKategoriRequest"}}}},"responses":{"200":{"description":""}},"tags":["Kategori"],"security":[{"bearer":[]}]},"get":{"operationId":"KategoriController_serach","parameters":[{"name":"id","required":false,"in":"query","schema":{"type":"number"}},{"name":"nama","required":false,"in":"query","schema":{"type":"string"}},{"name":"jenisbarang","required":false,"in":"query","schema":{"enum":["Barang","Jasa","Menu","Paket"],"type":"string"}},{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"size","required":false,"in":"query","example":10,"schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Kategori"],"security":[{"bearer":[]}]}},"/api/kategori/{kategori_id}":{"get":{"operationId":"KategoriController_get","summary":"","description":"Get Kategori data for given id","parameters":[{"name":"kategori_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Kategori"],"security":[{"bearer":[]}]},"put":{"operationId":"KategoriController_update","summary":"","description":"Update kategori record","parameters":[{"name":"kategori_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatekategoriRequest"}}}},"responses":{"200":{"description":""}},"tags":["Kategori"],"security":[{"bearer":[]}]},"delete":{"operationId":"KategoriController_remove","parameters":[{"name":"kategori_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Kategori"],"security":[{"bearer":[]}]}},"/api/subkategori":{"post":{"operationId":"SubkategoriController_create","summary":"","description":"Create a new subKategori Record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubkategoriRequest"}}}},"responses":{"200":{"description":""}},"tags":["Subkategori"],"security":[{"bearer":[]}]},"get":{"operationId":"SubkategoriController_serach","parameters":[{"name":"id","required":false,"in":"query","schema":{"type":"number"}},{"name":"nama","required":false,"in":"query","schema":{"type":"string"}},{"name":"jenisbarang","required":false,"in":"query","schema":{"enum":["Barang","Jasa","Menu","Paket"],"type":"string"}},{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"size","required":false,"in":"query","example":10,"schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Subkategori"],"security":[{"bearer":[]}]}},"/api/subkategori/{subkategori_id}":{"get":{"operationId":"SubkategoriController_get","summary":"","description":"Get Subkategori data for given id","parameters":[{"name":"subkategori_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Subkategori"],"security":[{"bearer":[]}]},"put":{"operationId":"SubkategoriController_update","summary":"","description":"Update kategori record","parameters":[{"name":"subkategori_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSubkategoriRequest"}}}},"responses":{"200":{"description":""}},"tags":["Subkategori"],"security":[{"bearer":[]}]},"delete":{"operationId":"SubkategoriController_remove","parameters":[{"name":"subkategori_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Subkategori"],"security":[{"bearer":[]}]}},"/api/satuan":{"post":{"operationId":"SatuanController_create","summary":"","description":"Create a new Kategori Record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSatuanRequest"}}}},"responses":{"200":{"description":""}},"tags":["Satuan"],"security":[{"bearer":[]}]},"get":{"operationId":"SatuanController_serach","parameters":[{"name":"id","required":false,"in":"query","schema":{"type":"number"}},{"name":"nama","required":false,"in":"query","schema":{"type":"string"}},{"name":"jenisbarang","required":false,"in":"query","schema":{"enum":["Barang","Jasa","Menu","Paket"],"type":"string"}},{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"size","required":false,"in":"query","example":10,"schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Satuan"],"security":[{"bearer":[]}]}},"/api/satuan/{satuan_id}":{"get":{"operationId":"SatuanController_get","summary":"","description":"Get Kategori data for given id","parameters":[{"name":"satuan_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Satuan"],"security":[{"bearer":[]}]},"put":{"operationId":"SatuanController_update","summary":"","description":"Update kategori record","parameters":[{"name":"satuan_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSatuanRequest"}}}},"responses":{"200":{"description":""}},"tags":["Satuan"],"security":[{"bearer":[]}]},"delete":{"operationId":"SatuanController_remove","parameters":[{"name":"satuan_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Satuan"],"security":[{"bearer":[]}]}},"/api/barang":{"post":{"operationId":"BarangController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBarangRequest"}}}},"responses":{"200":{"description":""}},"tags":["Barang"],"security":[{"bearer":[]}]},"get":{"operationId":"BarangController_search","parameters":[{"name":"barang_id","required":true,"in":"query","schema":{"type":"number"}},{"name":"size","required":true,"in":"query","schema":{"type":"number"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"nama","required":true,"in":"query","schema":{"type":"string"}},{"name":"kategori","required":true,"in":"query","schema":{"type":"string"}},{"name":"subkategori","required":true,"in":"query","schema":{"type":"string"}},{"name":"satuan","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Barang"],"security":[{"bearer":[]}]}},"/api/barang/{barang_id}":{"put":{"operationId":"BarangController_update","parameters":[{"name":"barang_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBarangRequest"}}}},"responses":{"200":{"description":""}},"tags":["Barang"],"security":[{"bearer":[]}]},"delete":{"operationId":"BarangController_remove","parameters":[{"name":"barang_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Barang"],"security":[{"bearer":[]}]},"get":{"operationId":"BarangController_detail","parameters":[{"name":"barang_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Barang"],"security":[{"bearer":[]}]}},"/api/suppliers":{"post":{"operationId":"SupplierController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSupplierRequest"}}}},"responses":{"200":{"description":""}},"tags":["Supplier"],"security":[{"bearer":[]}]},"get":{"operationId":"SupplierController_list","parameters":[{"name":"nama","required":false,"in":"query","schema":{"type":"string"}},{"name":"alamat","required":false,"in":"query","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","schema":{"type":"string"}},{"name":"no_hp","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"size","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Supplier"],"security":[{"bearer":[]}]}},"/api/suppliers/{contact_id}":{"get":{"operationId":"SupplierController_detail","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Supplier"],"security":[{"bearer":[]}]},"delete":{"operationId":"SupplierController_delete","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Supplier"],"security":[{"bearer":[]}]},"put":{"operationId":"SupplierController_update","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSupplierRequest"}}}},"responses":{"200":{"description":""}},"tags":["Supplier"],"security":[{"bearer":[]}]}},"/api/gudangs":{"post":{"operationId":"GudangController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGudangRequest"}}}},"responses":{"200":{"description":""}},"tags":["Gudang"],"security":[{"bearer":[]}]},"get":{"operationId":"GudangController_search","parameters":[{"name":"nama","required":true,"in":"query","schema":{"type":"string"}},{"name":"alamat","required":true,"in":"query","schema":{"type":"string"}},{"name":"keterangan","required":true,"in":"query","schema":{"type":"string"}},{"name":"is_active","required":true,"in":"query","schema":{"type":"boolean"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"size","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Gudang"],"security":[{"bearer":[]}]}},"/api/gudangs/{gudang_id}":{"get":{"operationId":"GudangController_detail","parameters":[{"name":"gudang_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Gudang"],"security":[{"bearer":[]}]},"delete":{"operationId":"GudangController_delete","parameters":[{"name":"gudang_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Gudang"],"security":[{"bearer":[]}]},"put":{"operationId":"GudangController_update","parameters":[{"name":"gudang_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGudangRequest"}}}},"responses":{"200":{"description":""}},"tags":["Gudang"],"security":[{"bearer":[]}]}},"/api/jenis_pelanggans":{"post":{"operationId":"JenisPelangganController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateJenisPelangganRequest"}}}},"responses":{"200":{"description":""}},"tags":["Jenis Pelanggan"],"security":[{"bearer":[]}]},"get":{"operationId":"JenisPelangganController_search","parameters":[{"name":"nama","required":false,"in":"query","example":"umum","schema":{"type":"string"}},{"name":"isDefault","required":false,"in":"query","example":false,"schema":{"type":"boolean"}},{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"size","required":false,"in":"query","example":10,"schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Jenis Pelanggan"],"security":[{"bearer":[]}]}},"/api/jenis_pelanggans/{jenisPelangganId}":{"get":{"operationId":"JenisPelangganController_detail","parameters":[{"name":"jenisPelangganId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Jenis Pelanggan"],"security":[{"bearer":[]}]},"delete":{"operationId":"JenisPelangganController_delete","parameters":[{"name":"jenisPelangganId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Jenis Pelanggan"],"security":[{"bearer":[]}]},"put":{"operationId":"JenisPelangganController_update","parameters":[{"name":"jenisPelangganId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateJenisPelangganRequest"}}}},"responses":{"200":{"description":""}},"tags":["Jenis Pelanggan"],"security":[{"bearer":[]}]}},"/api/pelanggans":{"post":{"operationId":"PelangganController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePelangganRequest"}}}},"responses":{"200":{"description":""}},"tags":["Pelanggan"],"security":[{"bearer":[]}]},"get":{"operationId":"PelangganController_search","parameters":[{"name":"nama","required":false,"in":"query","schema":{"type":"string"}},{"name":"alamat","required":false,"in":"query","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","schema":{"type":"string"}},{"name":"no_hp","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"size","required":false,"in":"query","schema":{"type":"number"}},{"name":"isCanCredit","required":false,"in":"query","schema":{"type":"boolean"}},{"name":"jenisPajak","required":false,"in":"query","schema":{"enum":["None","Inclusive","Exclusive"],"type":"string"}},{"name":"jenisPelanggan","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Pelanggan"],"security":[{"bearer":[]}]}},"/api/pelanggans/{contact_id}":{"get":{"operationId":"PelangganController_detail","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Pelanggan"],"security":[{"bearer":[]}]},"delete":{"operationId":"PelangganController_delete","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Pelanggan"],"security":[{"bearer":[]}]},"put":{"operationId":"PelangganController_update","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePelangganRequest"}}}},"responses":{"200":{"description":""}},"tags":["Pelanggan"],"security":[{"bearer":[]}]}},"/api/contacts":{"post":{"operationId":"ContactController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateContactRequest"}}}},"responses":{"200":{"description":""}},"tags":["Contact"],"security":[{"bearer":[]}]},"get":{"operationId":"ContactController_list","parameters":[{"name":"nama","required":false,"in":"query","schema":{"type":"string"}},{"name":"alamat","required":false,"in":"query","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","schema":{"type":"string"}},{"name":"no_hp","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"size","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Contact"],"security":[{"bearer":[]}]}},"/api/contacts/{contact_id}":{"get":{"operationId":"ContactController_detail","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Contact"],"security":[{"bearer":[]}]},"delete":{"operationId":"ContactController_delete","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Contact"],"security":[{"bearer":[]}]},"put":{"operationId":"ContactController_update","parameters":[{"name":"contact_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateContactRequest"}}}},"responses":{"200":{"description":""}},"tags":["Contact"],"security":[{"bearer":[]}]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/api/kas":{"post":{"operationId":"KasController_createKasAccount","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateKasDto"}}}},"responses":{"200":{"description":""}},"tags":["Kas"],"security":[{"bearer":[]}]},"get":{"operationId":"KasController_search","parameters":[{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"isActive","required":false,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["Kas"],"security":[{"bearer":[]}]}},"/api/kas/{kas_id}":{"put":{"operationId":"KasController_updateKasAccount","parameters":[{"name":"kas_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateKasDto"}}}},"responses":{"200":{"description":""}},"tags":["Kas"],"security":[{"bearer":[]}]},"get":{"operationId":"KasController_getById","parameters":[{"name":"kas_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Kas"],"security":[{"bearer":[]}]}}},"info":{"title":"System Toko","description":"The System Toko API","version":"1.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterUserRequestDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"nama":{"type":"string"},"alamat":{"type":"string"},"email":{"type":"string"},"no_hp":{"type":"string"}},"required":["username","password","nama"]},"AuthSignIn":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"CreateKategoriRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","enum":["Barang","Jasa","Menu","Paket"]},"nama":{"type":"string"}},"required":["jenis_barang","nama"]},"UpdatekategoriRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","enum":["Barang","Jasa","Menu","Paket"]},"id":{"type":"number"},"nama":{"type":"string"}},"required":["jenis_barang","id","nama"]},"CreateSubkategoriRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","enum":["Barang","Jasa","Menu","Paket"]},"nama":{"type":"string"}},"required":["jenis_barang","nama"]},"UpdateSubkategoriRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","enum":["Barang","Jasa","Menu","Paket"]},"id":{"type":"number"},"nama":{"type":"string"}},"required":["jenis_barang","id","nama"]},"CreateSatuanRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","enum":["Barang","Jasa","Menu","Paket"]},"nama":{"type":"string"}},"required":["jenis_barang","nama"]},"UpdateSatuanRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","enum":["Barang","Jasa","Menu","Paket"]},"id":{"type":"number"},"nama":{"type":"string"}},"required":["jenis_barang","id","nama"]},"CreateBarangRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","default":"Barang","enum":["Barang","Jasa","Menu","Paket"]},"nama":{"type":"string"},"satuan_id":{"type":"number"},"kategori_id":{"type":"number"},"subkategori_id":{"type":"number"},"hpp":{"type":"number","default":0},"harga_beli_akhir":{"type":"number","default":0},"harga_beli_3":{"type":"number","default":0},"harga_jual_1":{"type":"number","default":0},"harga_jual_2":{"type":"number","default":0},"harga_jual_3":{"type":"number","default":0},"barcode":{"type":"string"},"expired":{"format":"date-time","type":"string"},"diskon_rp":{"type":"number","default":0},"diskon_persen":{"type":"number","default":0},"jumlah_stok":{"type":"number","default":0},"jumlah_awal":{"type":"number","default":0},"keterangan":{"type":"string","default":""},"is_active":{"type":"boolean","default":true}},"required":["jenis_barang","nama","satuan_id","kategori_id","subkategori_id","hpp","harga_beli_akhir","harga_beli_3","harga_jual_1","harga_jual_2","harga_jual_3","barcode","diskon_rp","diskon_persen","jumlah_stok","jumlah_awal","keterangan","is_active"]},"UpdateBarangRequest":{"type":"object","properties":{"jenis_barang":{"type":"string","enum":["Barang","Jasa","Menu","Paket"]},"id":{"type":"number"},"nama":{"type":"string"},"satuan_id":{"type":"number"},"kategori_id":{"type":"number"},"subkategori_id":{"type":"number"},"hpp":{"type":"number"},"harga_beli_akhir":{"type":"number"},"harga_beli_3":{"type":"number"},"harga_jual_1":{"type":"number"},"harga_jual_2":{"type":"number"},"harga_jual_3":{"type":"number"},"barcode":{"type":"string"},"expired":{"format":"date-time","type":"string"},"diskon_rp":{"type":"number"},"diskon_persen":{"type":"number"},"jumlah_stok":{"type":"number"},"jumlah_awal":{"type":"number"},"keterangan":{"type":"string"},"is_active":{"type":"boolean"}},"required":["jenis_barang","id","nama","satuan_id","kategori_id","subkategori_id","hpp","harga_beli_akhir","harga_beli_3","harga_jual_1","harga_jual_2","harga_jual_3","barcode","diskon_rp","diskon_persen","jumlah_stok","jumlah_awal","keterangan","is_active"]},"CreateContactRequest":{"type":"object","properties":{"nama":{"type":"string"},"alamat":{"type":"string"},"email":{"type":"string"},"no_hp":{"type":"string"}},"required":["nama"]},"CreateSupplierRequest":{"type":"object","properties":{"contact_id":{"type":"number"},"contact":{"$ref":"#/components/schemas/CreateContactRequest"},"saldo_hutang":{"type":"number"},"max_hutang":{"type":"number"},"saldo_awal_hutang":{"type":"number"},"jatuh_tempo":{"type":"number"}}},"UpdateSupplierRequest":{"type":"object","properties":{"contact_id":{"type":"number"},"saldo_hutang":{"type":"number"},"max_hutang":{"type":"number"},"saldo_awal_hutang":{"type":"number"},"jatuh_tempo":{"type":"number"}},"required":["contact_id"]},"CreateGudangRequest":{"type":"object","properties":{"nama":{"type":"string"},"alamat":{"type":"string"},"keterangan":{"type":"string"},"is_active":{"type":"boolean","default":true}},"required":["nama","is_active"]},"UpdateGudangRequest":{"type":"object","properties":{"id":{"type":"number"},"nama":{"type":"string"},"alamat":{"type":"string"},"keterangan":{"type":"string"},"is_active":{"type":"boolean","default":true}},"required":["id","is_active"]},"CreateJenisPelangganRequest":{"type":"object","properties":{"nama":{"type":"string"}},"required":["nama"]},"UpdateJenisPelangganRequest":{"type":"object","properties":{"id":{"type":"number"},"nama":{"type":"string"},"isDefault":{"type":"boolean"}},"required":["id","nama","isDefault"]},"CreatePelangganRequest":{"type":"object","properties":{"contactId":{"type":"number"},"contact":{"$ref":"#/components/schemas/CreateContactRequest"},"jenisPelangganId":{"type":"number"},"isCanCredit":{"type":"boolean","default":false},"saldoPiutang":{"type":"number","default":0},"maxPiutang":{"type":"number","default":0},"limitHariPiutang":{"type":"number","default":0},"jatuhTempo":{"type":"number","default":0},"jenisPajak":{"type":"object","default":"None"}},"required":["jenisPelangganId"]},"UpdatePelangganRequest":{"type":"object","properties":{"contactId":{"type":"number"},"jenisPelangganId":{"type":"number"},"isCanCredit":{"type":"boolean","default":false},"saldoPiutang":{"type":"number","default":0},"maxPiutang":{"type":"number","default":0},"limitHariPiutang":{"type":"number","default":0},"jatuhTempo":{"type":"number","default":0},"jenisPajak":{"type":"object","default":"None"}}},"UpdateContactRequest":{"type":"object","properties":{"id":{"type":"number"},"nama":{"type":"string"},"alamat":{"type":"string"},"email":{"type":"string"},"no_hp":{"type":"string"}},"required":["id"]},"TypeKas":{"type":"string","enum":["01","02"]},"CreateKasDto":{"type":"object","properties":{"typeKas":{"$ref":"#/components/schemas/TypeKas"},"kodeAkunKartuKredit":{"type":"string"},"nama":{"type":"string"},"nomorRekening":{"type":"string"},"pemilik":{"type":"string"},"saldoKas":{"type":"number"},"keterangan":{"type":"string"},"isActive":{"type":"boolean"}},"required":["typeKas","kodeAkunKartuKredit","nama","nomorRekening","pemilik","saldoKas","keterangan","isActive"]},"UpdateKasDto":{"type":"object","properties":{"id":{"type":"number"},"typeKas":{"type":"string","enum":["01","02"]},"kodeAkunKartuKredit":{"type":"string"},"nama":{"type":"string"},"nomorRekening":{"type":"string"},"pemilik":{"type":"string"},"saldoKas":{"type":"number"},"keterangan":{"type":"string"},"isActive":{"type":"boolean"}},"required":["id","typeKas","kodeAkunKartuKredit","nama","nomorRekening","pemilik","saldoKas","keterangan","isActive"]}}}}